requirements.txt
Pillow>=9.0.0
numpy>=1.21.0
tkinterdnd2>=0.3.0 